+++ common/src/main/java/com/nb6868/onex/common/pojo/LoginByCodeReq.java
@@ -1,6 +1,5 @@
 package com.nb6868.onex.common.pojo;
 
-import com.nb6868.onex.common.validator.group.DefaultGroup;
 import com.nb6868.onex.common.validator.group.TenantGroup;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
@@ -13,11 +12,11 @@ import lombok.EqualsAndHashCode;
 public class LoginByCodeReq extends BaseReq {
 
     @Schema(description = "登录配置编码", example = "ADMIN_DINGTALK_CODE")
-    @NotEmpty(message = "登录类型不能为空", groups = {DefaultGroup.class})
+    // @NotEmpty(message = "登录类型不能为空", groups = {DefaultGroup.class})
     private String type;
 
     @Schema(description = "登录编码")
-    @NotEmpty(message = "登录编码不能为空", groups = {DefaultGroup.class})
+    @NotEmpty(message = "登录编码不能为空")
     private String code;
 
     @Schema(description = "租户编码")