+++ common/src/main/java/com/nb6868/onex/common/pojo/LoginByUsernamePasswordReq.java
@@ -1,6 +1,5 @@
 package com.nb6868.onex.common.pojo;
 
-import com.nb6868.onex.common.validator.group.DefaultGroup;
 import com.nb6868.onex.common.validator.group.TenantGroup;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
@@ -13,7 +12,7 @@ import lombok.EqualsAndHashCode;
 public class LoginByUsernamePasswordReq extends CaptchaReq {
 
     @Schema(description = "登录配置编码", example = "ADMIN_USERNAME_PASSWORD")
-    @NotEmpty(message = "请指定登录类型", groups = {DefaultGroup.class})
+    // @NotEmpty(message = "请指定登录类型", groups = {DefaultGroup.class})
     private String type;
 
     @Schema(description = "租户编码")
@@ -21,11 +20,15 @@ public class LoginByUsernamePasswordReq extends CaptchaReq {
     private String tenantCode;
 
     @Schema(description = "用户名")
-    @NotEmpty(message = "用户名不能为空", groups = {DefaultGroup.class})
+    @NotEmpty(message = "用户名不能为空")
     private String username;
 
-    @Schema(description = "密码(加密后)")
-    @NotEmpty(message = "密码不能为空", groups = {DefaultGroup.class})
+    @Schema(description = "密码(加密后,优先)")
+    // @NotEmpty(message = "密码不能为空", groups = {DefaultGroup.class})
     private String passwordEncrypted;
+
+    @Schema(description = "密码(明文)")
+    // @NotEmpty(message = "密码不能为空", groups = {DefaultGroup.class})
+    private String password;
 
 }